requirements.txt
fastapi==0.95.2
uvicorn==0.24.0
openai==0.28.1
langchain==0.0.351
chromadb==0.4.22
pydantic==1.10.13
python-dotenv==1.0.0
tiktoken==0.5.2
requests==2.31.0